Add ListContent class, ListType enum, and OptAList component
Added a new namespace `OptionA.Blazor.Blog` and within it, created a
new class `ListContent` that inherits from `Content`. This class
includes properties for `ListType`, a list of items, and an override
for `IsInvalid` to check if the list is empty.

Introduced a new enum `ListType` within the `OptionA.Blazor.Blog`
namespace to define the type of list, either `OrderedList` or
`UnorderedList`.

Created a new Razor component `OptAList.razor` within the
`OptionA.Blazor.Blog` namespace. This component dynamically renders
either an ordered list (`<ol>`) or an unordered list (`<ul>`) based
on the `ListType` property of the `Content`. Each list item includes
an input field and a remove button.

Added a code-behind file `OptAList.razor.cs` for the `OptAList`
component. This partial class includes a `Content` parameter of type
`ListContent` and an injected `IBlogDataProvider` to provide data
and attributes for the component.